# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is diamond play.",
        "title": "Diamond And Mine",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:7171",
    "basePath": "/",
    "paths": {
        "/diamond": {
            "post": {
                "description": "method to get some diamonds from a mine. If mine is empty, this method delete this mine.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get Diamonds from Mine.",
                "parameters": [
                    {
                        "description": "the name of the mine from which we want to extract diamonds",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.mineName"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.mineCount"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show the status of server.",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/mine": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show all mines.",
                "responses": {
                    "200": {
                        "description": "list of all mines",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repository.Mine"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add new mine.",
                "parameters": [
                    {
                        "description": "new mine model",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repository.Mine"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "gets all the diamonds from the mine and closes it",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Closes mine.",
                "parameters": [
                    {
                        "description": "the name of the mine to be closed",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.mineName"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.mineCount"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.mineCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
        "api.mineName": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "repository.Mine": {
            "type": "object",
            "properties": {
                "diamond_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}